requirements.txt
altgraph==0.17.4
certifi==2025.4.26
charset-normalizer==3.4.2
colorama==0.4.6
ffmpeg-python==0.2.0
filelock==3.13.1
fsspec==2024.6.1
future==1.0.0
idna==3.10
Jinja2==3.1.4
llvmlite==0.44.0
MarkupSafe==2.1.5
more-itertools==10.7.0
mpmath==1.3.0
networkx==3.3
numba==0.61.2
numpy==2.1.2
openai-whisper @ git+https://github.com/openai/whisper.git@dd985ac4b90cafeef8712f2998d62c59c3e62d22
packaging==25.0
pefile==2023.2.7
pillow==11.0.0
pyasn1==0.6.1
pyinstaller==6.13.0
pyinstaller-hooks-contrib==2025.4
pywin32-ctypes==0.2.3
regex==2024.11.6
requests==2.32.3
rsa==4.9.1
setuptools==70.2.0
sympy==1.13.3
tiktoken==0.9.0
torch==2.7.0+cu118
torchaudio==2.7.0+cu118
torchvision==0.22.0+cu118
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.4.0
